backend/requirements.txt
# Django Core
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-otp==1.5.4
django-two-factor-auth==1.16.0
qrcode==7.4.2
django-axes==6.4.0
phonenumbers==8.13.47

# Background Tasks
celery==5.3.6
redis==5.0.3
flower==2.0.1
eventlet==0.36.1
# django-celery-beat==2.6.0  # Commented out due to Django 5.1 compatibility issues

# Development
python-dotenv==1.0.1
dj-database-url==2.1.0
